make things quieter, and mark 4.3.6 stable on x86

+++ dev-php/php/php-4.3.5.ebuild
@@ -1,6 +1,6 @@
 # Copyright 1999-2004 Gentoo Technologies, Inc.
 # Distributed under the terms of the GNU General Public License v2
-# $Header: /var/cvsroot/gentoo-x86/dev-php/php/php-4.3.5.ebuild,v 1.3 2004/03/30 19:28:58 avenj Exp $
+# $Header: /var/cvsroot/gentoo-x86/dev-php/php/php-4.3.5.ebuild,v 1.4 2004/05/04 23:23:24 robbat2 Exp $
 
 PHPSAPI="cli"
 inherit php-sapi eutils
@@ -46,7 +46,7 @@ pkg_postinst() {
 	einfo "This is a CLI only build."
 	einfo "You cannot use it on a webserver."
 
-	if [ "`md5sum ${ROOT}/root/.pearrc`" = "f0243f51b2457bc545158cf066e4e7a2  ${ROOT}/root/.pearrc" ]; then
+	if [ -f "${ROOT}/root/.pearrc" -a "`md5sum ${ROOT}/root/.pearrc`" = "f0243f51b2457bc545158cf066e4e7a2  ${ROOT}/root/.pearrc" ]; then
 		einfo "Cleaning up an old PEAR install glitch"
 		mv ${ROOT}/root/.pearrc ${ROOT}/root/.pearrc.`date +%Y%m%d%H%M%S`
 	fi
